refactor(geo): consolidate shared types and validation
