+++ apps/dashboard/src/components/geo/traffic-hero.tsx
@@ -13,6 +13,7 @@ import {
   trafficSparklineDays,
   trafficVisitDelta,
 } from "@notra/geo-core/utils/ai-traffic";
+import { todayIsoDate } from "@notra/geo-core/utils/day-label";
 import { useState } from "react";
 
 import { EChartsAreaChart } from "@/components/evilcharts/charts/echarts-area-chart";
@@ -29,7 +30,6 @@ import {
   toggleTrafficTrendKey,
   trafficTrendProviderTypeKey,
 } from "@/utils/ai-traffic-trend";
-import { todayIsoDate } from "@/utils/analytics-charts";
 import { seriesColors } from "@/utils/chart-colors";
 import { engineIconHtml } from "@/utils/engine-icon-html";
 import { formatChartInteger } from "@/utils/geo-charts";
