+++ apps/dashboard/src/constants/geo-directions.ts
@@ -7,6 +7,7 @@ import type {
   GeoJourney,
   GeoTrafficPage,
 } from "@notra/geo-core/types/geo";
+import { formatDayLabel } from "@notra/geo-core/utils/day-label";
 
 import { CHART_PRIMARY_COLOR, CHART_SECONDARY_COLOR } from "@/constants/charts";
 import type { ChartConfig } from "@/types/charts";
@@ -19,7 +20,6 @@ import type {
   GeoDirectionTab,
   GeoDirectionTone,
 } from "@/types/geo-directions";
-import { formatDayLabel } from "@/utils/analytics-charts";
 import { seriesColors } from "@/utils/chart-colors";
 import { chartKey } from "@/utils/chart-keys";
 import {
