+++ apps/dashboard/src/constants/onboarding.ts
@@ -11,13 +11,6 @@ export const ONBOARDING_VISIBLE_SUGGESTIONS = 3;
 export const ONBOARDING_SUGGESTION_SKELETON_ROWS = [0, 1, 2, 3] as const;
 export const ONBOARDING_FIELD_CLASS = "h-11 rounded-xl px-3.5";
 
-export {
-  ONBOARDING_WEBSITE_IPV4_REGEX,
-  ONBOARDING_WEBSITE_IPV6_REGEX,
-  ONBOARDING_WEBSITE_PREFIX_REGEX,
-  ONBOARDING_WEBSITE_TOP_LEVEL_DOMAIN_REGEX,
-} from "@notra/geo-core/constants/website-url";
-
 export const ONBOARDING_HEARD_ABOUT_NOTRA_SOURCES = [
   "x",
   "github",
