--- apps/dashboard/src/lib/geo-ingest/generation.ts
@@ -1,5 +1,0 @@
-/**
- * Moved to `@notra/geo-core/geo/ingest` so the public API can issue and revoke
- * the same tracking tokens. Re-exported here to keep existing import paths.
- */
-export { getGeoIngestTokenGeneration } from "@notra/geo-core/geo/ingest";
