--- apps/dashboard/src/types/generations/tracking.ts
@@ -1,4 +1,0 @@
-export type {
-  ActiveGeneration,
-  GenerationResult,
-} from "@notra/geo-core/types/generation-tracking";
